Added crawler.parse package and its tests
Parse_Base class was added in parse_base.py to crawler.parse
subpackage. parse subpackage is expected to parse fetched data
into text, html or sections(container). HTML(webpage), PDF,
DOCX and PPTX needs to be supported.

If its difficult to parse fetched, it could just be converted to
text by implementing .to_text(). If its possible, data can be
parsed to html which can be easily worked on due to availability
of parsers for html.

If theres a lot to be extracted from fetch object(fetched data),
it may be better to create sections manually and add the data
to their matadata. This is reccommended but may be diffiult to
archieve as it may require lot of inspection of fetch object
data using parser library.

Sections cannot only be created manually but can be automated.
utility.Text already is capable of breaking text into sections
based on max_section_size. Other data about text can be used
to create sections including new lines used for creating sections
for representing lines.

HTML tags can also be used based on whats already is known from
the parser. If fetched data could be converted to HTML then it
would be easy to parse it and create sections from that HTML. That
would free the need to haing to learn and use lot parsing libraries
for different files and sources.

Converting to sections may require different modules. Converting text
to sections may also require its own module to analyse and recreate
meaning from the text.