Added implementation of .is_fetch_valid() method
.is_fetch_valid() method of Fetch_Base method is based on whole source(url)
is wont always be valid. Not all urls with html code will end with .html
or similar. Mimetypes is not directly used on whole url(source) but
on the path part.

So if path part point to file then its mimetype is guessed. If its
none then url point to webpage which is likely be rich in HTML
code. https://example.com is expected to be parsed with HTML_Parse
other parse for files ending with '.com' extension.

The trick used could be applied to files in web. But most of the time,
urls pointing to files will end in that file extension e.g https://example.com/file.pdf
which mimetype can easily guess the type. But if url contain get request data
then it would be hard. But getting the path part could help or removing part
with with url encoded data.

It certanly works but I cant guarantee it.