Updated get_urls_from_text() function
Regex to match urls was updated and re.findall() was used instead of
re.search. The regex wasnt tested on variety of urls, it could be unpredictable
or match invalid urls or fail to match the whole url.

Extracting urls from html is better than using get_urls_from_text() unless pattern
that is good in matching urls was used. I tried to modify ther regex pattern and thats
whats I was able to produse. The pattern does not take into account domain name which
means it will match urls without domain names but having path part.

It may also fail to match other parts of url including data in url. It may need to be
tested if its able to perform well.