Ranamed 3 functions of pattern get_*() to create_*()
get_fetch_object() was rename to create_fetch_object(), the same applies
to create_parse_object() and create_crawl_object. The rename came as this
functions create objects other than retrieving it somewhere. Starting with
'get' would food users that it runs in constant time while it has to create
objects which could take seconds.

staring with create makes it clear that object will be created.